requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0

# SegFormer dependencies
transformers>=4.20.0
timm>=0.6.0
mmcv-full>=1.4.0  # Optional, for advanced features

# OpenPose (reuse from original)
einops>=0.3.0

# Utilities
tqdm>=4.62.0
pathlib

